Move data tables to separate file
